Get download_url to point at pulp content app
The issue here is that the pulp_ansible ContentVersionSerializer
doesn't include the distro base_path in the url, presumably
because it was always empty before?

Current pulp-content-app provides collection artifacts at
urls like:

    http://localhost:24816/pulp/content/foo-distro-base-path/alikins.collection_inspect.0.0.182.tar.gz

The 'foo-distro-base-path' is the base path for the
'foo-distro' which is the distro associated with the
'foo' Repository.

Current pulp_ansible leaves out the 'foo-distro-base-path' even
if the pulp config settings for CONTENT_ORIGIN and CONTENT_PATH_PREFIX
are tweaked.

So galaxy_ng.app.api.v3.serializers.GalaxyNGCollectionVersionSerializer
is a workaround. Should probably get fixed in pulp_ansible.

At this point, 'ansible-galaxy collection publish' works with the
relocated URL endpoints.

'ansible-galaxy collection install' also works.