Add CollectionVersion serial that takes repo/dist in download_url
Use a GalaxyNGCollectionVersionSerializer to be used with
the CollectionVersionsViewset.

This serializer includes the Collections repo/dist base_path
in the download_url path.

The issue here is that the pulp_ansible ConllectionVersionSerializer
doesn't include the distro base_path in the url, presumably
because it was always empty before?

Current pulp-content-app provides collection artifacts at
urls like:

    http://localhost:24816/pulp/content/foo-distro-base-path/alikins.collection_inspect.0.0.182.tar.gz

The 'foo-distro-base-path' is the base path for the
'foo-distro' which is the distro associated with the
'foo' Repository.

Current pulp_ansible leaves out the 'foo-distro-base-path' even
if the pulp config settings for CONTENT_ORIGIN and CONTENT_PATH_PREFIX
are tweaked.

So galaxy_ng.app.api.v3.serializers.GalaxyNGCollectionVersionSerializer
is a workaround. Should probably get fixed in pulp_ansible.

TODO: This serializer could be removed if pulp_ansible
handles it in the future.
See https://pulp.plan.io/issues/6510