Use PulpTemporaryFile
Find import CollectionVersions vis Task CreatedResources

Since PulpTemporaryFile doesn't have a reference to the
Artifact we can't lookup the created CollectionVersion by
the Artifact id.

So lookup the current task, find it's CreatedResources, and
grabbed the CollectionVersions.

Note: In theory, since there is a list of imported CollectionVersions
now, a single import could in the future created multiple CVs.

This requires:
- pulpcore >= 3.6.0
- pulp_ansible > 0.2.0b15

Issue: #330
Issue: #333