On changes to golden repo, trigger all synclist tasks
In the CollectionMoveViewSet, if either of the repositories
involved is the 'golden' repo, then enqueue a
'curate_all_synclist_repository' task that start
tasks that curate all synclist repositories based on
the synclist associated with it.

ie, when the main repo changes, kick off tasks to update
all the repos based on it.

Issue: #428