Fix import task errors about ansible.CollectionImport
The pkey of galaxy's CollectionImport (galaxy_collectionimport table)
is task_id. But the 'task_id' is a OneToOne to a
pulp_ansible.app.models.CollectionImport now instead of just
a plain UUID.

So, get the right pulp CollectionImport and use it for
task_id when creating a galaxy_ng.app.models.CollectionImport.

Closes-Issue: AAH-99