Avoid the IntegrityError when trying to duplicate existing name.
When trying to create a namespace with existing name
the IntegrityError was raised and should be a 400 with detail.

This needed to be solved on viewset before the creation of
dependent instances of different models because they all
use the same naming.

It was tried to fix that on the NamespaceSerializer.validate_name
but that happens after the Viewset.create is invoked.

Issue: AAH-132