Add bump2version config to setup.cfg
Use bump2version to increment the 'version' string wherever
it is needed. The canonical source of truth for the version
is 'bump2version.current_version' in setup.cfg.

To update version, change that and run 'bump2version'.

Our to just do a 'patch' version bump, can use
'make dev/bumpversion-patch' that will rev the 'patch'
level and update all the files that use it.

- Convience 'make' targets for bumpversion
- Borrow x.y.z-{release} handling from pulp bumpversion
- Setup current version in setup.py, bumpversion will handle
- Add requirements.shipit.txt for reqs needed to build and ship a release