Make token and password obfuscated on the API docs for /sync/config
On /sync/config the fields password and token was being exposed
as plain text, according to pulp devs this is not an issue
as they decided to leave it exposed because RBAC + https
takes care of it.

In our end, for the matters of UX we decided to obfuscate those
fields and make them write_only so it is not possible to
access password or token via API but only set it.

> **NOTE**: this password is not related to proxy, this is
the password used to authenticate to the remote.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1920613
Issue: AAH-282