On publish to 'myrepo', try 'inbound-myrepo', then 'myrepo'
If pushing to /content/inbound-mynamespace/, then push to
inbound-mynamespace.

if pushing to /content/mynamespace/, try inbound-mynamespace, then
mynamespace.

If a 'inbound-published' repo/dist is created, then a publish
to /content/published/ will push to 'inbound-published' instead
of 'staging'. If auto-approve is enabled, the collection will
be auto moved from inbound-published to published.

Issue: AAH-558