Stop running the synclist curate all task
Previously, any changes to the upstream / golden
repo would cause the 'curate_all_synclist_repository'
task to run. This updates _all_ synclists to reflect
the upstream repo and that was too much.

In combination with changes to private automation hub,
the synclist curation is no longer needed so this change
stops it from running.

Issue: AHA-1556