Role import logging (#1959)
Add a custom logger to emit logging events to a model. This architecture mimics that of pulp_ansible's collection upload logging. The model has a json column to store all log events as a list of dictionaries. The imports are also listable and filterable. The serializer for imports list is polymorphic if the "id" parameter is given.

* Import log model for roles and prettier logging.
* Rename migration file.
* Handle logging emits for non-import tasks.
* Default task_id
* No user not fatal.
* Switch to modelviewset and allow listing.
* Prettify the import logs.

No-Issue

Signed-off-by: James Tanner <tanner.jc@gmail.com>