fix: Fresh installs must get the correct URL for rh-certified
Add /content/published to the default URL when setting up the
default remotes.

This has been fix on #2032 but that PR not backported.

This PR changes the original migration, so fresh installs gets the new URL
and existing systems are not affected

This must be backported to 4.9.z

No-Issue
Related: AAH-2836