Integration tests should sleep a lot less. (#2225)
Most of the time, integration tests are sleeping and waiting to poll an endpoint. That poll time
is defined in a constant which can be overridden via env vars, but there were still places in the test
code that we not using the constant. This change fixes all those places where possible and injects
very small sleep times via environment variables in the various run_integration.sh scripts.

No-Issue

Signed-off-by: James Tanner <tanner.jc@gmail.com>