rewrite _ui/v2/users/ and add more test coverage (#2229)
Unpin connection to v1 user view and serializer and instead implement from scratch. 
Add full support for group/team/organization POSTs and PUTs with the user. 
Integration tests should cover most aspects of the viewset and various expectations.

Passwords are not required, but must be valid.
Emails are not required, but must be valid.

No-Issue

Signed-off-by: James Tanner <tanner.jc@gmail.com>