Incorporate ruff linter into CI pipeline
The ruff tool is a linting and formatting tool, that replaces
most of the functionality of the flake8 including most popular plugins,
isort and black.

This commit adds ruff into development and CI pipeline with minimal
configuration. This enables only basic checks that will be expanded in
future aiming to replace flake8 with plugins and isort.

Fixed warnings:

11      F401    [*] unused-import
 3      F841    [*] unused-variable
 2      F811    [*] redefined-while-unused
 1      E741    [ ] ambiguous-variable-name